Moved utility-fucntions
