+++ src/app/consolidated-goals/consolidated-goals.component.ts
@@ -4,7 +4,7 @@ import {MatPaginator} from '@angular/material/paginator';
 import {MatSort, Sort} from '@angular/material/sort';
 import {MatTableDataSource} from '@angular/material/table';
 import {GoalSummary} from '../generated-data-api';
-import {reformatYYYYMMDD} from '../../utility-functions';
+import {reformatYYYYMMDD} from '../util/utility-functions';
 
 @Component({
   selector: 'app-consolidated-goals',
