+++ src/app/services/data.service.ts
@@ -56,7 +56,7 @@ import {
   reformatYYYYMMDD,
   formatWotResult,
   getWotLineChartAnnotationsObject
-} from '../../utility-functions';
+} from '../util/utility-functions';
 import {patchTsGetExpandoInitializer} from '@angular/compiler-cli/ngcc/src/packages/patch_ts_expando_initializer';
 import {ChartDataSets, ChartPoint} from 'chart.js';
 import * as moment from 'moment';
