+++ src/app/services/goals-data-service.service.ts
@@ -8,7 +8,7 @@ import {environment} from '../../environments/environment';
 import {GoalLists, GoalTarget, MccGoal, MccObservation} from '../generated-data-api';
 
 import {TargetValue} from '../datamodel/targetvalue';
-import {formatGoalTargetValue} from '../../utility-functions';
+import {formatGoalTargetValue} from '../util/utility-functions';
 import {VitalSignsTableData} from '../datamodel/vitalSigns';
 import {EgfrTableData} from '../datamodel/egfr';
 import {UacrTableData} from '../datamodel/uacr';
