+++ src/app/uacr/uacr.component.ts
@@ -5,7 +5,7 @@ import {DataService} from '../services/data.service';
 import {MatPaginator} from '@angular/material/paginator';
 import {MatSort, Sort} from '@angular/material/sort';
 import {UacrTableData} from '../datamodel/uacr';
-import {formatUacrResult, reformatYYYYMMDD} from '../../utility-functions';
+import {formatUacrResult, reformatYYYYMMDD} from '../util/utility-functions';
 import {MatTableDataSource} from '@angular/material/table';
 
 @Component({
